notification-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
aio-pika==9.3.0
aiormq==6.7.7
python-dotenv==1.0.0
